+++ R/view_xifti_surface.R
@@ -2,7 +2,7 @@
 #'
 #' See \code{\link{view_xifti_surface}} for details.
 #'
-#' @param xifti The \code{"xifti"}
+#' @inheritParams xifti_Param
 #' @param surfL Left surface
 #' @param surfR Right surface
 #' @param hemisphere Hemisphere
@@ -293,7 +293,9 @@ view_xifti_surface.color <- function(
     if (color_mode=="qualitative") {
       # For .dlabel files, use the included labels metadata colors.
       if ((!is.null(xifti_meta$cifti$intent) && xifti_meta$cifti$intent==3007)) {
-        if (length(idx) > 1) { message("Color labels from first requested column will be used.") }
+        if (length(unique(xifti_meta$cifti$labels[idx])) > 1) {
+          message("Color labels from first requested column will be used.")
+        }
         labs <- xifti_meta$cifti$labels[[idx[1]]]
         if (is.null(colors)) {
           pal_base <- data.frame(
@@ -349,7 +351,7 @@ view_xifti_surface.color <- function(
 
   } else {
     pal <- pal_base <- NULL
-    # Will become NA_COLOR
+    # Will become NA_color
     values <- list(left=matrix(0), right=matrix(0))
   }
 
@@ -572,6 +574,11 @@ view_xifti_surface.draw_mesh <- function(
 #'  \code{TRUE} borders will be colored in black; provide the name of a different
 #'  color to use that instead. If \code{FALSE} or \code{NULL} (default), do
 #'  not draw borders.
+#' @param shadows Number from 0 (maximum added lighting) to 1 (no added 
+#'  lighting) to control the darkness and extent of shadowing on the 3D surface. 
+#'  Default: \code{1}. Shadows help render the shape of the surface, but can 
+#'  be distracting if interpretation of the data depends on small differences in
+#'  brightness along the color scale.
 #' @return If a png or html file(s) were written, the names of the files for
 #'  each index (and color legend if applicable) will be returned. Otherwise,
 #'  the widget itself is returned if a widget was used, and the rgl object IDs
@@ -590,11 +597,12 @@ view_xifti_surface <- function(
   view=c("both", "lateral", "medial"), widget=NULL,
   title=NULL, slider_title="Index",
   fname=FALSE, fname_suffix=c("names", "idx"), legend_fname="[fname]_legend",
-  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL, 
+  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
   cex.title=NULL, text_color="black", bg=NULL,
-  borders=FALSE, alpha=1.0, edge_color=NULL, vertex_color=NULL, vertex_size=0,
-  material=NULL,
+  NA_color="white", borders=FALSE, alpha=1.0,
+  edge_color=NULL, vertex_color=NULL, vertex_size=0,
+  material=NULL, shadows=1,
   width=NULL, height=NULL, zoom=NULL
   ) {
 
@@ -989,7 +997,6 @@ view_xifti_surface <- function(
 
   # Former arguments that are now internal
   render_rgl <- TRUE # for debugging
-  NA_COLOR <- "white" # surfaces without data will be white
   close_after_save <- TRUE # always close after saving
   widget_idx_warn <- TRUE # always warn if widget might take a while to load
 
@@ -1208,6 +1215,38 @@ view_xifti_surface <- function(
       rgl::open3d()
       if (is.null(bg)) { bg <- "white" }
       rgl::bg3d(color=bg)
+      stopifnot(is.numeric(shadows) && shadows>=0 && shadows<=1)
+      if (shadows < 1) {
+        more_light <- 1- shadows
+        # Set maximum to 25% brightness.
+        more_light <- more_light/4
+        # Convert to a grayscale hex color.
+        more_light <- round(more_light*255)
+        more_light <- as.hexmode(more_light)
+        more_light <- paste0("#", paste0(rep(more_light, 3), collapse=""))
+        # Lights from top and bottom
+        rgl::light3d(
+          theta=0, phi=80,
+          viewpoint.rel = TRUE,
+          diffuse = more_light,
+        )
+        rgl::light3d(
+          theta=0, phi=-80,
+          viewpoint.rel = TRUE,
+          diffuse = more_light,
+        )
+        # Lights from sides
+        rgl::light3d(
+          theta=80, phi=0,
+          viewpoint.rel = TRUE,
+          diffuse = more_light,
+        )
+        rgl::light3d(
+          theta=-80, phi=0,
+          viewpoint.rel = TRUE,
+          diffuse = more_light,
+        )
+      }
       rgl::par3d(windowRect = 40 + c(0, 0, round(all_panels_width), round(all_panels_height)))
       Sys.sleep(1) #https://stackoverflow.com/questions/58546011/how-to-draw-to-the-full-window-in-rgl
       if (use_title && length(hemisphere) > 1) {
@@ -1296,9 +1335,9 @@ view_xifti_surface <- function(
       } else {
         # Get the color.
         if (any_colors) {
-          mesh_color <- c(NA_COLOR, as.character(pal$color))[color_vals[[h]][,jj] + 1]
+          mesh_color <- c(NA_color, as.character(pal$color))[color_vals[[h]][,jj] + 1]
         } else {
-          mesh_color <- rep(NA_COLOR, nrow(switch(h, left=surfL, right=surfR)$vertices))
+          mesh_color <- rep(NA_color, nrow(switch(h, left=surfL, right=surfR)$vertices))
         }
         # Draw border.
         if (!is.null(borders) && color_mode=="qualitative") {
@@ -1447,6 +1486,9 @@ view_xifti_surface <- function(
     }
 
     # [TO DO]: Adjust sizing
+    # Looked into htmlwidgets::sizingPolicy but it doesn't affect how the legend
+    #   scales with the plot width, which can't be controlled within RStudio
+    #   pane.
     out <- rgl::rglwidget(
       height=round(all_panels_height/1.1),
       width=round(all_panels_width/1.1)
@@ -1537,10 +1579,12 @@ view_cifti_surface <- function(
   view=c("both", "lateral", "medial"), widget=NULL,
   title=NULL, slider_title="Index",
   fname=FALSE, fname_suffix=c("names", "idx"), legend_fname="[fname]_legend",
-  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL, 
+  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
   cex.title=NULL, text_color="black", bg=NULL,
-  borders=FALSE, alpha=1.0, edge_color=NULL, vertex_color=NULL, vertex_size=0,
+  NA_color="white", borders=FALSE, alpha=1.0,
+  edge_color=NULL, vertex_color=NULL, vertex_size=0,
+  material=NULL, shadows=1,
   width=NULL, height=NULL, zoom=NULL){
 
   view_xifti_surface(
@@ -1551,10 +1595,12 @@ view_cifti_surface <- function(
     view=view, widget=widget,
     title=title, slider_title=slider_title,
     fname=fname, fname_suffix=fname_suffix, legend_fname=legend_fname,
-    legend_ncol=legend_ncol, legend_alllevels=legend_alllevels, legend_embed=legend_embed, 
+    legend_ncol=legend_ncol, legend_alllevels=legend_alllevels, legend_embed=legend_embed,
     digits=digits, scientific=scientific,
     cex.title=cex.title, text_color=text_color, bg=bg,
-    borders=borders, alpha=alpha, edge_color=edge_color, vertex_color=vertex_color, vertex_size=vertex_size,
+    NA_color=NA_color, borders=borders, alpha=alpha,
+    edge_color=edge_color, vertex_color=vertex_color, vertex_size=vertex_size,
+    material=material, shadows=shadows,
     width=width, height=height, zoom=zoom
   )
 }
@@ -1565,14 +1611,16 @@ viewCIfTI_surface <- function(
   xifti=NULL, surfL=NULL, surfR=NULL,
   color_mode="auto", zlim=NULL, colors=NULL,
   idx=NULL, hemisphere=NULL,
-  together=NULL, together_ncol=together_ncol, together_title=NULL,
+  together=NULL, together_ncol=NULL, together_title=NULL,
   view=c("both", "lateral", "medial"), widget=NULL,
   title=NULL, slider_title="Index",
   fname=FALSE, fname_suffix=c("names", "idx"), legend_fname="[fname]_legend",
-  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL, 
+  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
   cex.title=NULL, text_color="black", bg=NULL,
-  borders=FALSE, alpha=1.0, edge_color=NULL, vertex_color=NULL, vertex_size=0,
+  NA_color="white", borders=FALSE, alpha=1.0,
+  edge_color=NULL, vertex_color=NULL, vertex_size=0,
+  material=NULL, shadows=1,
   width=NULL, height=NULL, zoom=NULL){
 
   view_xifti_surface(
@@ -1583,10 +1631,12 @@ viewCIfTI_surface <- function(
     view=view, widget=widget,
     title=title, slider_title=slider_title,
     fname=fname, fname_suffix=fname_suffix, legend_fname=legend_fname,
-    legend_ncol=legend_ncol, legend_alllevels=legend_alllevels, legend_embed=legend_embed, 
+    legend_ncol=legend_ncol, legend_alllevels=legend_alllevels, legend_embed=legend_embed,
     digits=digits, scientific=scientific,
     cex.title=cex.title, text_color=text_color, bg=bg,
-    borders=borders, alpha=alpha, edge_color=edge_color, vertex_color=vertex_color, vertex_size=vertex_size,
+    NA_color=NA_color, borders=borders, alpha=alpha,
+    edge_color=edge_color, vertex_color=vertex_color, vertex_size=vertex_size,
+    material=material, shadows=shadows,
     width=width, height=height, zoom=zoom
   )
 }
@@ -1597,14 +1647,16 @@ viewcii_surface <- function(
   xifti=NULL, surfL=NULL, surfR=NULL,
   color_mode="auto", zlim=NULL, colors=NULL,
   idx=NULL, hemisphere=NULL,
-  together=NULL, together_ncol=together_ncol, together_title=NULL,
+  together=NULL, together_ncol=NULL, together_title=NULL,
   view=c("both", "lateral", "medial"), widget=NULL,
   title=NULL, slider_title="Index",
   fname=FALSE, fname_suffix=c("names", "idx"), legend_fname="[fname]_legend",
-  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL, 
+  legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
   cex.title=NULL, text_color="black", bg=NULL,
-  borders=FALSE, alpha=1.0, edge_color=NULL, vertex_color=NULL, vertex_size=0,
+  NA_color="white", borders=FALSE, alpha=1.0,
+  edge_color=NULL, vertex_color=NULL, vertex_size=0,
+  material=NULL, shadows=1,
   width=NULL, height=NULL, zoom=NULL){
 
   view_xifti_surface(
@@ -1615,10 +1667,12 @@ viewcii_surface <- function(
     view=view, widget=widget,
     title=title, slider_title=slider_title,
     fname=fname, fname_suffix=fname_suffix, legend_fname=legend_fname,
-    legend_ncol=legend_ncol, legend_alllevels=legend_alllevels, legend_embed=legend_embed, 
+    legend_ncol=legend_ncol, legend_alllevels=legend_alllevels, legend_embed=legend_embed,
     digits=digits, scientific=scientific,
     cex.title=cex.title, text_color=text_color, bg=bg,
-    borders=borders, alpha=alpha, edge_color=edge_color, vertex_color=vertex_color, vertex_size=vertex_size,
+    NA_color=NA_color, borders=borders, alpha=alpha,
+    edge_color=edge_color, vertex_color=vertex_color, vertex_size=vertex_size,
+    material=material, shadows=shadows,
     width=width, height=height, zoom=zoom
   )
 }