+++ R/view_xifti_volume.R
@@ -186,7 +186,7 @@
 #'  The vertical positioning is controlled by \code{ypos.title}, and the font
 #'  size is controlled by \code{cex.title}.
 #' @param crop Crop the slice subplots to the subcortical structures, instead of
-#'  showing the full anatomical image? Default: \code{TRUE}. 
+#'  showing the full anatomical image? Default: \code{TRUE}.
 #'  Ignored if \code{widget}.
 #' @param text_color Color for text in title and colorbar legend. Default:
 #'  \code{"white"}. If \code{"white"}, will use black instead for the color
@@ -231,7 +231,7 @@ view_xifti_volume <- function(
   legend_fname="[fname]_legend",
   legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
-  cex.title=NULL, ypos.title=0, xpos.title=0, 
+  cex.title=NULL, ypos.title=0, xpos.title=0,
   orientation_labels=TRUE, crop=TRUE,
   text_color="white", bg=NULL, width=NULL, height=NULL, ...) {
 
@@ -624,7 +624,9 @@ view_xifti_volume <- function(
     if (color_mode=="qualitative") {
       # For .dlabel files, use the included labels metadata colors.
       if ((!is.null(xifti$meta$cifti$intent) && xifti$meta$cifti$intent==3007)) {
-        if (length(idx) > 1) { message("Color labels from first requested column will be used.") }
+        if (length(unique(xifti$meta$cifti$labels[idx])) > 1) {
+          message("Color labels from first requested column will be used.")
+        }
         labs <- xifti$meta$cifti$labels[[idx[1]]]
         if (is.null(colors)) {
           pal_base <- data.frame(
@@ -886,6 +888,8 @@ view_xifti_volume <- function(
   if (!together && makePDF) { pdf(fname, width=width, height=height) }
 
   zlim <- sort(zlim)[c(1, length(zlim))]
+  # Problem if zlim has range of 0
+  if (zlim[1] == zlim[2]) { zlim[2] <- zlim[2] + 1e-8 }
 
   for (jj in seq(length(idx))) {
     this_idx <- idx[jj]
@@ -898,7 +902,10 @@ view_xifti_volume <- function(
     img_labels@.Data <- labs_bs
     img_labels@.Data[labs_bs==0] <- NA
 
-    if (crop) {
+    # Avoid error if all NA
+    if (all(is.na(img_overlay@.Data))) { img_overlay@.Data[] <- 0}
+
+    if (crop && (!all(is.na(img_overlay@.Data)))) {
       crop_x <- which(apply(!is.na(img_overlay@.Data), 1, any))
       crop_x <- seq(min(crop_x), max(crop_x))
       crop_y <- which(apply(!is.na(img_overlay@.Data), 2, any))
@@ -921,6 +928,11 @@ view_xifti_volume <- function(
       img2 <- img[slices,crop_y,crop_z]
       img_overlay2 <- img_overlay[slices,crop_y,crop_z,drop=FALSE]
     } else { stop() }
+
+    # Replace `NA` with out-of-zlim value, because `NA` has problems sometimes
+    # and if it's all `NA`, it's definitely a problem. Out-of-zlim values
+    # become transparent.
+    img_overlay2[is.na(img_overlay2)] <- min(zlim) - 1
 
     oro.nifti::overlay(
       x=img2, y=img_overlay2, plane=plane,
@@ -1013,7 +1025,7 @@ view_xifti_volume <- function(
           width = (legend_ncol) * cleg_h_per_row * cleg_w_factor
         )
       }
-      if (close_after_save) { dev.off() }
+      if (close_after_save && !((together && !comp_dummy) || together_leg)) { dev.off() }
     }
   } else if (use_cleg && !comp_dummy && !together_leg && isFALSE(fname)) {
     print(cleg)
@@ -1151,7 +1163,7 @@ view_cifti_volume <- function(
   legend_fname="[fname]_legend",
   legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
-  cex.title=NULL, ypos.title=0, xpos.title=0, 
+  cex.title=NULL, ypos.title=0, xpos.title=0,
   orientation_labels=TRUE, crop=TRUE,
   text_color="white", bg=NULL, width=NULL, height=NULL, ...) {
 
@@ -1189,7 +1201,7 @@ viewCIfTI_volume <- function(
   legend_fname="[fname]_legend",
   legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
-  cex.title=NULL, ypos.title=0, xpos.title=0, 
+  cex.title=NULL, ypos.title=0, xpos.title=0,
   orientation_labels=TRUE, crop=TRUE,
   text_color="white", bg=NULL, width=NULL, height=NULL, ...) {
 
@@ -1227,7 +1239,7 @@ viewcii_volume <- function(
   legend_fname="[fname]_legend",
   legend_ncol=NULL, legend_alllevels=FALSE, legend_embed=NULL,
   digits=NULL, scientific=NA,
-  cex.title=NULL, ypos.title=0, xpos.title=0, 
+  cex.title=NULL, ypos.title=0, xpos.title=0,
   orientation_labels=TRUE, crop=TRUE,
   text_color="white", bg=NULL, width=NULL, height=NULL, ...) {